examples/bad-apple/pyproject.toml
[project]
name = "bad-apple"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "yt-dlp",
    "python-ffmpeg",
    "pympress",
    "pycairo",
    "pygobject",
]
